notebooks/requirement.txt
streamlit
pandas
plotly
requests
python-dotenv
supabase-py